OK. This solves the weird rotation issue with iPhone SE, running 14.
It seems to be an OS bug, related to a bottom toolbar, using autolayout. If you attach the toolbar to the safe area (not the superview, using AL, it seems to cause a hang, on rotate.
What I did, was add an "intervening" UIView, between the toolbar and the safe area, and that seems to spike the hang.